requirements.txt
matplotlib
pandas
mplfinance
beautifulsoup4
requests
numpy
Pillow
